+++ src/calendar-base/calendar-base.tsx
@@ -2,7 +2,7 @@ import { css } from "atomico";
 import { CalendarMonthContext } from "../calendar-month/CalendarMonthContext.js";
 import { reset } from "../utils/styles.js";
 import type { DaysOfWeek } from "../utils/utils.js";
-import { PlainDate } from "../utils/PlainDate.js";
+import type { PlainDate } from "../utils/temporal.js";
 import type { DateWindow } from "../utils/DateWindow.js";
 
 type CalendarBaseProps = {
