+++ src/calendar-base/useCalendarBase.ts
@@ -1,7 +1,6 @@
 import { useState, useEvent, useHost } from "atomico";
-import { type Duration, DateWindow } from "../utils/DateWindow.js";
-import { PlainDate } from "../utils/PlainDate.js";
-import { PlainYearMonth } from "../utils/PlainYearMonth.js";
+import { DateWindow } from "../utils/DateWindow.js";
+import { type PlainDate, PlainYearMonth } from "../utils/temporal.js";
 import { useDateProp, useDateFormatter } from "../utils/hooks.js";
 import { today } from "../utils/utils.js";
